+++ src/components/pages/splash_screen.js
